move api utils to separate module
